remove enum member annotations
"Members defined within an enum class should not include explicit type annotations. Type checkers should infer a literal type for all members. A type checker should report an error if a type annotation is used for an enum member because this type will be incorrect and misleading to readers of the code"

https://typing.readthedocs.io/en/latest/spec/enums.html#defining-members